+++ src/dataStructures/bloomFilter/types/index.ts
@@ -1,8 +1,0 @@
-export interface BloomFilterInterface {
-  insert: (item: string) => void;
-  contain: (item: string) => boolean;
-  hash1: (item: string) => number;
-  hash2: (item: string) => number;
-  hash3: (item: string) => number;
-  getHashValues: (item: string) => number[];
-}